WTF crash error
There was an error that crashed the app when the gather checkbox was shown on the login page. This only happened on the first run of the app, when the "google login" part was shown (permissions for google fit access). I Don't know why the error happened. I gave up on it and just removed the component. Perhaps someone might want to spend some time figuring this out....